Move and rename the MemoryAccesses class to MemoryAccessTracker.
